refactor: move operator and operator datagrid code to separate files
